[python3] Fixes the majority of failing tests (c.f. #157)

+++ python/bob/ap/test/test_energy.py
@@ -3,16 +3,16 @@
 # Elie Khoury <Elie.Khoury@idiap.ch>
 #
 # Copyright (C) 2011-2013 Idiap Research Institute, Martigny, Switzerland
-# 
+#
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
 # the Free Software Foundation, version 3 of the License.
-# 
+#
 # This program is distributed in the hope that it will be useful,
 # but WITHOUT ANY WARRANTY; without even the implied warranty of
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 # GNU General Public License for more details.
-# 
+#
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
@@ -25,7 +25,7 @@ import math
 import time
 
 #############################################################################
-# Tests blitz-based extrapolation implementation with values returned 
+# Tests blitz-based extrapolation implementation with values returned
 #############################################################################
 
 ########################## Values used for the computation ##################
@@ -33,7 +33,7 @@ eps = 1e-3
 
 #############################################################################
 numpy.set_printoptions(precision=2, threshold=numpy.nan, linewidth=200)
-  
+
 def _read(filename):
   """Read video.FrameContainer containing preprocessed frames"""
 
@@ -48,7 +48,7 @@ def _read(filename):
 
 def compare(v1, v2, width):
   return abs(v1-v2) <= width
-  
+
 def mel_python(f):
   import math
   return 2595.0*math.log10(1.+f/700.0)
@@ -60,13 +60,13 @@ def sig_norm(win_length, frame, flag):
   gain = 0.0
   for i in range(win_length):
     gain = gain + frame[i] * frame[i]
-  
+
   ENERGY_FLOOR = 1.0
   if gain < ENERGY_FLOOR:
     gain = math.log(ENERGY_FLOOR)
   else:
     gain = math.log(gain)
-    
+
   if(flag and gain != 0.0):
     for i in range(win_length):
       frame[i] = frame[i] / gain
@@ -80,7 +80,7 @@ def pre_emphasis(frame, win_length, a):
   else:
     for i in range(win_length - 1, 0, -1):
       frame[i] = frame[i] - a * frame[i - 1]
-    frame[0] = (1. - a) * frame[0]  
+    frame[0] = (1. - a) * frame[0]
   return frame
 
 def hamming_window(vector, hamming_kernel, win_length):
@@ -94,7 +94,7 @@ def log_filter_bank(x, n_filters, p_index, win_size):
   for i in range(0, win_size / 2 + 1):
     re = complex_[i].real
     im = complex_[i].imag
-    x[i] = math.sqrt(re * re + im * im) 
+    x[i] = math.sqrt(re * re + im * im)
   filters = log_triangular_bank(x, n_filters, p_index)
   return filters, x
 
@@ -102,8 +102,8 @@ def log_triangular_bank(data, n_filters, p_index):
   a = 1.0 / (p_index[1:n_filters+2] - p_index[0:n_filters+1] + 1)
   vec1 =  list(numpy.arange(p_index[i], p_index[i + 1]) for i in range(0, n_filters))
   vec2 =  list(numpy.arange(p_index[i+1], p_index[i + 2] + 1) for i in range(0, n_filters))
-  res_ = numpy.array([(numpy.sum(data[vec1[i]]*(1.0 - a [i]* (p_index[i + 1]-(vec1[i])))) + 
-          numpy.sum(data[vec2[i]] * (1.0 - a[i+1] * ( (vec2[i]) - p_index[i + 1])))) 
+  res_ = numpy.array([(numpy.sum(data[vec1[i]]*(1.0 - a [i]* (p_index[i + 1]-(vec1[i])))) +
+          numpy.sum(data[vec2[i]] * (1.0 - a[i+1] * ( (vec2[i]) - p_index[i + 1]))))
           for i in range(0, n_filters)])
   FBANK_OUT_FLOOR = 1.0
   filters = numpy.log(numpy.where(res_ < FBANK_OUT_FLOOR, FBANK_OUT_FLOOR, res_))
@@ -120,7 +120,7 @@ def dct_transform(filters, n_filters, dct_kernel, n_ceps, dct_norm):
   for i in range(1, n_ceps + 1):
     ceps[i - 1] = numpy.sum(filters[vec - 1] * dct_kernel[i - 1][0:n_filters])
     ceps[i - 1] = ceps[i - 1] * dct_coeff
-    
+
   return ceps
 
 
@@ -129,7 +129,7 @@ def energy_computation(obj, rate_wavsample, win_length_ms, win_shift_ms, n_filte
   #########################
   ## Initialisation part ##
   #########################
-  
+
   c = bob.ap.Energy(rate_wavsample[0], win_length_ms, win_shift_ms)
 
   #ct = bob.ap.TestCeps(c)
@@ -142,14 +142,14 @@ def energy_computation(obj, rate_wavsample, win_length_ms, win_shift_ms, n_filte
   win_size = int (2.0 ** math.ceil(math.log(win_length) / math.log(2)))
   m = int (math.log(win_size) / math.log(2))
 
-  # Hamming initialisation 
+  # Hamming initialisation
   cst = 2 * math.pi / (win_length - 1.0)
   hamming_kernel = numpy.zeros(win_length)
-  
+
   for i in range(win_length):
     hamming_kernel[i] = (0.54 - 0.46 * math.cos(i * cst))
 
-  # Compute cut-off frequencies 
+  # Compute cut-off frequencies
   p_index = numpy.array(numpy.zeros(n_filters + 2), dtype=numpy.int16)
   if(mel_scale):
     # Mel scale
@@ -169,8 +169,8 @@ def energy_computation(obj, rate_wavsample, win_length_ms, win_shift_ms, n_filte
       p_index[i] = int (round((win_size / (sf * 1.0) * f)))
 
   #Cosine transform initialisation
-  dct_kernel = [ [ 0 for i in range(n_filters) ] for j in range(n_ceps) ] 
-  
+  dct_kernel = [ [ 0 for i in range(n_filters) ] for j in range(n_ceps) ]
+
   for i in range(1, n_ceps + 1):
     for j in range(1, n_filters + 1):
       dct_kernel[i - 1][j - 1] = math.cos(math.pi * i * (j - 0.5) / n_filters)
@@ -178,37 +178,37 @@ def energy_computation(obj, rate_wavsample, win_length_ms, win_shift_ms, n_filte
   ######################################
   ### End of the Initialisation part ###
   ######################################
-  
+
   ######################################
   ###          Core code             ###
   ######################################
 
   data_size = data.shape[0]
-  n_frames = 1 + (data_size - win_length) / win_shift
- 
+  n_frames = int(1 + (data_size - win_length) / win_shift)
+
   # create features set
 
-  params = [ 0 for j in range(n_frames) ] 
-   
+  params = [ 0 for j in range(n_frames) ]
+
   # compute cepstral coefficients
   delta = 0
   for i in range(n_frames):
     # create a frame
     frame = numpy.zeros(win_size, dtype=numpy.float64)
     som = 0.0
-    vec = numpy.arange(win_length)  
+    vec = numpy.arange(win_length)
     frame[vec] = data[vec + i * win_shift]
     som = numpy.sum(frame)
     som = som / win_size
-    frame = frame - som  
-    
+    frame = frame - som
+
 
     energy = sig_norm(win_length, frame, False)
-    
+
     params[i] = energy
 
   data = numpy.array(params)
-  
+
   return data
 
 
@@ -218,21 +218,21 @@ def energy_comparison_run(obj, rate_wavsample, win_length_ms, win_shift_ms, n_fi
 
   #ct = bob.ap.TestCeps(c)
   A = c(rate_wavsample[1])
- 
-  B = energy_computation(obj, rate_wavsample, win_length_ms, win_shift_ms, n_filters, n_ceps, dct_norm, 
+
+  B = energy_computation(obj, rate_wavsample, win_length_ms, win_shift_ms, n_filters, n_ceps, dct_norm,
         f_min, f_max, delta_win, pre_emphasis_coef, mel_scale, with_energy, with_delta, with_delta_delta)
 
   diff=numpy.sum(numpy.sum((A-B)*(A-B)))
   obj.assertAlmostEqual(diff, 0., 7, "Error in Ceps Analysis")
 
-##################### Unit Tests ##################  
+##################### Unit Tests ##################
 class EnergyTest(unittest.TestCase):
   """Test the Energy extraction"""
 
-  def test_energy(self): 
+  def test_energy(self):
     import pkg_resources
     rate_wavsample = _read(pkg_resources.resource_filename(__name__, os.path.join('data', 'sample.wav')))
-    
+
     win_length_ms = 20
     win_shift_ms = 10
     n_filters = 24
@@ -242,10 +242,10 @@ class EnergyTest(unittest.TestCase):
     delta_win = 2
     pre_emphasis_coef = 0.97
     dct_norm = True
-    mel_scale = True 
+    mel_scale = True
     with_energy = True
     with_delta = True
     with_delta_delta = True
-    
+
     energy_comparison_run(self,rate_wavsample, win_length_ms, win_shift_ms, n_filters, n_ceps, dct_norm, f_min, f_max, delta_win,
                                pre_emphasis_coef, mel_scale, with_energy, with_delta, with_delta_delta)
